Rmc 45 backend do test for endpoints

+++ backend/src/main/java/com/rmc/pejo/endpoints/ReminderEndPoint.java
@@ -1,7 +1,10 @@
 package com.rmc.pejo.endpoints;
 
 import com.rmc.pejo.entity.Reminder;
+import com.rmc.pejo.exceptions.ResourceNotFoundException;
+import com.rmc.pejo.service.PetService;
 import com.rmc.pejo.service.ReminderService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
@@ -12,12 +15,11 @@ import java.util.Set;
 @CrossOrigin
 @RestController
 @RequestMapping("reminders")
+@RequiredArgsConstructor
 public class ReminderEndPoint {
     private final ReminderService service;
 
-    public ReminderEndPoint(ReminderService service) {
-        this.service = service;
-    }
+    private final PetService petService;
 
     @PostMapping
     public Reminder save(@Valid @RequestBody Reminder reminder) {
@@ -30,8 +32,8 @@ public class ReminderEndPoint {
     }
 
     @GetMapping("{id}")
-    public Optional<Reminder> get(@PathVariable Long id) {
-        return service.get(id);
+    public Reminder get(@PathVariable Long id) {
+        return service.get(id).orElseThrow(ResourceNotFoundException::new);
     }
 
     @GetMapping("/first3ByDate")
@@ -46,12 +48,14 @@ public class ReminderEndPoint {
 
     @DeleteMapping("{id}")
     public String deleteById(@PathVariable("id") Long id) {
+        if (service.get(id).isEmpty()) throw new ResourceNotFoundException();
         service.delete(id);
         return "Deleted successfully";
     }
 
     @GetMapping("pet/{petId}")
     public Set<Reminder> getRemindersByPetId(@PathVariable("petId") Long petId) {
+        if (petService.get(petId).isEmpty()) throw new ResourceNotFoundException();
         return service.getRemindersByPetId(petId);
     }
 }
